refactor blend and add blend by forecast horizon as well
